+++ test/unit-disabled/_generated_google.tf
@@ -2,7 +2,8 @@
 // TERRAMATE: originated from generate_hcl block on /test/terramate_google.tm.hcl
 
 variable "gcp_project" {
-  description = "(Required) The ID of the project in which the resource belongs."
+  default     = null
+  description = "(Optional) The ID of the project in which the resource belongs."
   type        = string
 }
 variable "gcp_region" {
@@ -12,13 +13,22 @@ variable "gcp_region" {
 }
 variable "gcp_org_domain" {
   default     = null
-  description = "(optional) The domain of the organization test projects should be created in."
+  description = "(Optional) The domain of the organization test projects should be created in."
+  type        = string
+}
+variable "gcp_billing_account" {
+  default     = null
+  description = "(Optional) The billing account to use when creating projects."
   type        = string
 }
 terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
+      version = "~> 4.0"
+    }
+    google-beta = {
+      source  = "hashicorp/google-beta"
       version = "~> 4.0"
     }
     random = {
@@ -31,6 +41,10 @@ provider "google" {
   project = var.gcp_project
   region  = var.gcp_region
 }
+provider "google-beta" {
+  project = var.gcp_project
+  region  = var.gcp_region
+}
 provider "random" {
 }
 resource "random_id" "suffix" {
@@ -38,16 +52,24 @@ resource "random_id" "suffix" {
   count       = local.test_name != "unit-disabled" ? 1 : 0
 }
 data "google_project" "project" {
+  count      = var.gcp_project != null ? 1 : 0
+  project_id = var.gcp_project
 }
 data "google_organization" "org" {
-  count  = local.org_domain != null ? 1 : 0
-  domain = local.org_domain
+  count  = var.gcp_org_domain != null ? 1 : 0
+  domain = var.gcp_org_domain
+}
+data "google_billing_account" "billing_account" {
+  billing_account = var.gcp_billing_account
+  count           = var.gcp_billing_account != null ? 1 : 0
 }
 locals {
-  org_domain     = var.gcp_org_domain
-  org_id         = try(data.google_organization.org[0].name, null)
-  project_id     = var.gcp_project
-  project_number = data.google_project.project.number
-  random_suffix  = try(random_id.suffix[0].hex, "disabled")
-  test_name      = "unit-disabled"
+  billing_account       = try(data.google_billing_account.billing_account[0].id, "no-billing-account")
+  directory_customer_id = try(data.google_organization.org[0].directory_customer_id, "no-directoy-customer")
+  org_domain            = try(data.google_organization.org[0].domain, "no-domain.mineiros.io")
+  org_id                = try(data.google_organization.org[0].org_id, null)
+  project_id            = try(data.google_project.project[0].project_id, "no-project")
+  project_number        = try(data.google_project.project[0].number, null)
+  random_suffix         = try(random_id.suffix[0].hex, "disabled")
+  test_name             = "unit-disabled"
 }