+++ test/unit-disabled/main.tf
@@ -3,9 +3,22 @@ module "test" {
 
   module_enabled = false
 
+  project = "unit-disabled-${local.random_suffix}"
   # add all required arguments
 
-  project = "your-project-id"
+  # add all optional arguments that create additional/extended resources
+}
+
+module "test2" {
+  source = "../.."
+
+  module_enabled = false
+
+  project = "unit-disabled-${local.random_suffix}"
+  # add all required arguments
+
+  authoritative = false
+  members       = ["user:member@example.com"]
 
   # add all optional arguments that create additional/extended resources
 }